+++ rs/canister/impl/src/lib.rs
@@ -1,3 +1,5 @@
+use querystring::QueryParams;
+
 mod email_sender;
 mod env;
 mod guards;
@@ -10,6 +12,13 @@ mod state;
 mod updates;
 
 type Hash = [u8; 32];
+
+fn get_query_param_value(params: &QueryParams, key: &str) -> Option<String> {
+    params
+        .iter()
+        .find(|(k, _)| *k == key)
+        .map(|(_, v)| v.to_string())
+}
 
 #[cfg(test)]
 mod generate_candid_file {