+++ rs/libraries/test_utils/src/lib.rs
@@ -1,9 +1,11 @@
+use magic_links::{MagicLink, SignedMagicLink};
 use rand::rngs::StdRng;
-use rand::SeedableRng;
+use rand::{thread_rng, SeedableRng};
 use rsa::pkcs1::LineEnding;
 use rsa::pkcs8::EncodePublicKey;
 use rsa::RsaPrivateKey;
-use sign_in_with_email_canister::{InitArgs, InitOrUpgradeArgs};
+use sign_in_with_email_canister::{Delegation, InitArgs, InitOrUpgradeArgs, TimestampNanos};
+use utils::ValidatedEmail;
 
 pub const TEST_SALT: [u8; 32] = [1; 32];
 pub const EMAIL_SENDER_RSA_SEED: [u8; 32] = [2; 32];
@@ -15,22 +17,43 @@ pub fn default_init_args() -> InitOrUpgradeArgs {
     })
 }
 
-pub fn rsa_private_key() -> RsaPrivateKey {
+pub fn generate_magic_link(
+    email: &str,
+    session_key: Vec<u8>,
+    created: TimestampNanos,
+    expiration: TimestampNanos,
+) -> SignedMagicLink {
+    let seed = utils::calculate_seed(
+        TEST_SALT,
+        &ValidatedEmail::try_from(email.to_string()).unwrap(),
+    );
+    let delegation = Delegation {
+        pubkey: session_key,
+        expiration,
+    };
+    let magic_link = MagicLink::new(seed, delegation, created);
+    let private_key = rsa_private_key();
+    let encrypted = magic_link.encrypt(private_key.to_public_key(), &mut thread_rng());
+
+    encrypted.sign(email_sender_rsa_private_key())
+}
+
+fn rsa_private_key() -> RsaPrivateKey {
     generate_rsa_private_key_from_seed(TEST_SALT)
 }
 
-pub fn email_sender_rsa_private_key() -> RsaPrivateKey {
+fn email_sender_rsa_private_key() -> RsaPrivateKey {
     generate_rsa_private_key_from_seed(EMAIL_SENDER_RSA_SEED)
 }
 
-pub fn email_sender_public_key_pem() -> String {
+fn email_sender_public_key_pem() -> String {
     email_sender_rsa_private_key()
         .to_public_key()
         .to_public_key_pem(LineEnding::LF)
         .unwrap()
 }
 
-pub fn generate_rsa_private_key_from_seed(seed: [u8; 32]) -> RsaPrivateKey {
+fn generate_rsa_private_key_from_seed(seed: [u8; 32]) -> RsaPrivateKey {
     let mut rng = StdRng::from_seed(seed);
     RsaPrivateKey::new(&mut rng, 2048).unwrap()
 }