+++ rs/canister/CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add an explicit endpoint for handling magic links ([#33](https://github.com/open-chat-labs/ic-sign-in-with-email/pull/33))
 
+### Fixed
+
+- Fix template updater lambda function ([#36](https://github.com/open-chat-labs/ic-sign-in-with-email/pull/36))
+
 ## [[0.7.0](https://github.com/open-chat-labs/ic-sign-in-with-email/releases/tag/v0.7.0)] - 2024-05-09
 
 ### Changed