+++ rs/canister/CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Remove dependency on `dfx-core` ([#37](https://github.com/open-chat-labs/ic-sign-in-with-email/pull/37))
 
+### Fixed
+
+- Fix template updater lambda function ([#36](https://github.com/open-chat-labs/ic-sign-in-with-email/pull/36))
+
 ## [[0.7.0](https://github.com/open-chat-labs/ic-sign-in-with-email/releases/tag/v0.7.0)] - 2024-05-09
 
 ### Changed