+++ rs/canister/CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add an explicit endpoint for handling magic links ([#33](https://github.com/open-chat-labs/ic-sign-in-with-email/pull/33))
 
+### Changed
+
+- Remove dependency on `dfx-core` ([#37](https://github.com/open-chat-labs/ic-sign-in-with-email/pull/37))
+
 ### Fixed
 
 - Fix template updater lambda function ([#36](https://github.com/open-chat-labs/ic-sign-in-with-email/pull/36))