+++ rs/canister/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+## [[0.9.0](https://github.com/open-chat-labs/ic-sign-in-with-email/releases/tag/v0.9.0)] - 2024-05-23
+
 ### Added 
 
 - Add an explicit endpoint for handling magic links ([#33](https://github.com/open-chat-labs/ic-sign-in-with-email/pull/33))